- Paired down github workflow files, adding nightly tests if a new commit has not yet been tested - Modernized gen_HICAR_dir

+++ helpers/gen_HICAR_dir.sh
@@ -4,9 +4,46 @@
 # i.e. :
 #./gen_HICAR_dir.sh ./Model_runs/ /home/user/HICAR/
 
+# check if the user supplied the correct number of arguments
+if [ "$#" -ne 2 ]; then
+	echo
+	echo "Usage: $0 <parent_directory> <HICAR_repo_directory>"
+	echo
+	exit 1
+fi
+
 # check if the directory at path $1 exists
 if [ ! -d "$1" ]; then
 	mkdir -p $1
+fi
+
+# check if the directory at path $2 exists
+if [ ! -d "$2" ]; then
+	echo
+	echo "Error: HICAR repo directory $2 does not exist."
+	echo
+	exit 1
+else [ -d "$2/bin" ]; then
+	echo
+	echo "Error: HICAR repo directory $2 does not contain a bin/ directory."
+	echo
+	exit 1
+fi
+
+# check for any files in $2/bin starting with HICAR
+found_hicar=false
+for _f in "$2"/bin/HICAR*; do
+	if [ -e "$_f" ]; then
+		found_hicar=true
+		break
+	fi
+done
+
+if [ "$found_hicar" = false ]; then
+	echo
+	echo "Error: HICAR repo directory $2 does not contain a HICAR executable in bin/"
+	echo
+	exit 1
 fi
 
 # Use readlink -f or realpath to get absolute paths
@@ -54,6 +91,15 @@ if [ ! -f ./NoahmpTable.TBL ]; then
 	echo $HICAR_dir/run
 	echo 'to ./input'
 
+	# check if the .TBL files exist in the HICAR repo run directory, if not,
+	# ask the user to re-run the configure script to download them
+	if [ ! -f $HICAR_dir/run/NoahmpTable.TBL ]; then
+		echo
+		echo 'Error: .TBL files not found in the HICAR repo run directory.'
+		echo 'Please re-run the cmake configuration step to download them.'
+		echo
+		exit 1
+	fi
 	cp $HICAR_dir/run/*.TBL ./
 fi
 
@@ -69,8 +115,35 @@ if [ ! -d ./rrtmg_support -o ! -d ./mp_support ]; then
 	echo 'https://github.com/NCAR/icar_supporting_files.git'
 	echo 
 
+	# check if the physics support folders exist in the HICAR repo run directory, if not,
+	# ask the user to re-run the configure script to download them
+	if [ ! -d $HICAR_dir/run/rrtmg_support ]; then
+		echo
+		echo 'Error: rrtmg_support directory not found in the HICAR repo run directory.'
+		echo 'Please re-run the cmake configuration step to download them.'
+		echo
+		exit 1
+	fi
+
+	if [ ! -d $HICAR_dir/run/rrtmgp_support ]; then
+		echo
+		echo 'Error: rrtmgp_support directory not found in the HICAR repo run directory.'
+		echo 'Please re-run the cmake configuration step to download them.'
+		echo
+		exit 1
+	fi
+
+	if [ ! -d $HICAR_dir/run/mp_support ]; then
+		echo
+		echo 'Error: mp_support directory not found in the HICAR repo run directory.'
+		echo 'Please re-run the cmake configuration step to download them.'
+		echo
+		exit 1
+	fi
+
 	cd $HICAR_dir
 	cp -r run/rrtmg_support $parent_dir/HICAR/input
+	cp -r run/rrtmgp_support $parent_dir/HICAR/input
 	cp -r run/mp_support $parent_dir/HICAR/input
 
 fi
